+++ src/data/yolo_dataset.cpp
@@ -56,7 +56,7 @@ DatasetSample YOLODataset::get_sample(size_t index) const {
     sample.orig_height = H;
     sample.orig_width = W;
 
-    sample.mask = Tensor::zeros({1, H, W});
+    sample.mask = Tensor::zeros({1, static_cast<int64_t>(H), static_cast<int64_t>(W)});
     float* mask_data = sample.mask->data();
 
     // Find label txt file with matching stem
